lambda/data_fetcher/requirements.txt
praw
SQLAlchemy
psycopg2-binary
python-dotenv
pydantic
